Refactor Python client to a separate module
